feat: Add SnapdealSort component for PDF and CSV processing
- Introduce SnapdealSort component to handle PDF and CSV file uploads for Snapdeal data extraction.
- Implement SKU, quantity, and company extraction logic tailored for Snapdeal PDF format.
- Enhance user interface with file upload feedback and dynamic processing status updates.
- Integrate SnapdealSort into extract-sku.js, replacing previous inline logic for improved modularity and maintainability.
